gtest changes

+++ socketcan_bridge/include/socketcan_bridge/socketcan_signal.hpp
@@ -58,7 +58,7 @@ class SocketCANSignal
     uint16_t start_bit_;
     std::string topic_name_;
 
-    float value_;
+    double value_;
 };
 
 }  // namespace socketcan_bridge
